dbtag: add a naive db schema migration system
For further features, the schema may need to be changed. And so
incremental changes must be supported.

Track current schema version in a single-cell table "version". Have an
UPGRADES dictionary listing SQL statements to execute to go from version
N to N+1.

"Empty db" and "no schema versioning yet" are considered version 0, and
"tags with schema versioning" is version 1.