dbtag: allow multithread open for fsops
File tree copy/move occur in a dedicated thread to avoid blocking the UI
but sqlite3 lib rejects by default db operations from another thread.
Since we use a modal dialog while the file operation is in progress,
multiple threads can access the db but not simulteneously. Thus, we can
use sqlite3 without a problem. Add a bool to enable this.

Another way would be to perform file operations in the main thread but
this requires significant changes, so we prefer the other way.