ui: when scrolling to some files, try to thumbnail them in priority
If a dir has 5000 files and the user scrolls straight to the end of list, we
want to favor those last files in thumbnail queue instead of doing the whole
dir in normal order.
Implement on-demand reordering, and perform it in model's data() method.
This assumes data() will be called preferably when the item is displayed on
screen rather than for all items.

OrderedDict seems an easy choice that meets our needs:
- appending at end of queue
- delete in arbitrary position (not known in advance) of queue
- delete and prepend an item of queue